+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## 1.1.0
 
-- Added RegistryModule.remove()
-- Added SearchResult:getFirst()
+- Added `RegistryModule.remove()`
+- Added `RegistryModule.searchFrom()`
+- Added `SearchResult:getFirst()`
 
-- Added `exclude` parameter in SearchResult:forEach() callback
+- Added `exclude` parameter in `SearchResult:forEach()` callback
+-
 
 ## 1.0.0